privacy-radar/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
pillow==10.4.0
lxml[html_clean]==5.3.0
pydantic==2.9.2
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
python-dotenv==1.0.1
